Make the LoginParser.py module import relative to  the project base.
The idea here is that all code execution -- tests and otherwise -- would take
place from the project's base directory, eliminating the need for sys.path
manipulation shenanigans.

With this change, running "pytest" from the base directory does not give me
import errors.  (The test doesn't pass, but that is a separate issue.)